// src/app/enums/service-request-change-payment-mode.enum.ts
export enum PaymentModes {
    Annual = 1,
    Semiannual = 2,
    Quarterly = 3,
    Monthly = 4
}
export enum PremiumTypes {
    ChangePlannedPremium = 1,
    ChangePremiumModeOtherThanMonthly = 2,
    ChangePremiumModeMonthly = 3
}
export enum IsPayorDifferentThanPolicyOwner{
    Yes = 1,
    No = 2
}
export enum AccountTypes{
    Checking = 1,
    Savings = 2
}